changed a bunch of files and got rendering working correctly
