+++ frontend/src/App.js
@@ -1,62 +1,40 @@
 import React, { Component } from 'react';
 import './App.css'
-import Navbar from './Navbar'
-// import Dashboard from './Dashboard'
-import LeadsTable from './tables/LeadsTable'
-// import EmployeeLeaderboard from './tables/EmployeeLeaderboardTable'
-// import ReferrerLeaderboard from './tables/ReferrerLeaderboardTable'
-import LodgementsTable from './tables/LodgementsTable'
-import ApprovalsTable from './tables/ApprovalsTable'
-import SettlementsTable from './tables/SettlementsTable'
-// import CRM from './tables/CrmTable'
-// import Form from './Form'
-// import Login from './Login'
+import Dashboard from './components/Dashboard'
+import EmployeeLeaderboard from './components/EmployeeLeaderboard'
+import ReferrerLeaderboard from './components/ReferrerLeaderboardTable'
+import LeadsTable from './components/LeadsTable'
+import LodgementsTable from './components/LodgementsTable'
+import ApprovalsTable from './components/ApprovalsTable'
+import SettlementsTable from './components/SettlementsTable'
+import CRM from './components/CrmTable'
+import Form from './components/Form'
+import Login from './components/Login'
+import { BrowserRouter, Route } from 'react-router-dom'
 
 
-class mainPage extends Component {
+class App extends Component {
   render() {
-
-    // const { pathname } = window.location;
-    // const params = pathname.substr(1);
-
-    // if (params === 'dashboard') {
-    //   return <Dashboard />
-    // } else if (params === 'leads') {
-    //   return <LeadsTable />
-    // } else if (params === 'leads/new') {
-    //   return <Form />
-    // }  else if (params === 'leads/employeeleaderboard') {
-    //   return <EmployeeLeaderboard />
-    // } else if (params === 'leads/referrerleaderboard') {
-    //   return <ReferrerLeaderboard />
-    // } else if (params === 'lodgements') {
-    //   return <LodgementsTable />
-    // } else if (params === 'approvals') {
-    //   return <ApprovalsTable />
-    // } else if (params === 'settlements') {
-    //   return <SettlementsTable />
-    // } else if (params === 'crm') {
-    //   return <CRM />
-    // } else if (params === 'login') {
-    //   return <Login />
-    // }
-
     return (
-      <div>
-
-        <Navbar />
         
-        <div class="all-tables">
-          <LeadsTable />
-          <LodgementsTable />
-          <ApprovalsTable />
-          <SettlementsTable />
-        </div>
+        <BrowserRouter>
+            <div>
+                <Route path='/dashboard' component={Dashboard} />
+                <Route exact path='/leads' component={LeadsTable} />
+                <Route path='/leads/:id' component={LeadsTable} />
+                <Route path='/leads/new' component={Form} />
+                <Route path='/leads/employeeleaderboard' component={EmployeeLeaderboard} />
+                <Route path='/leads/referrerleaderboard' component={ReferrerLeaderboard} />
+                <Route exact path='/lodgements' component={LodgementsTable} />
+                <Route exact path='/approvals' component={ApprovalsTable} />
+                <Route exact path='/settlements' component={SettlementsTable} />
+                <Route path='/crm' component={CRM} />
+                <Route path='/login' component={Login} />
+            </div>
+        </BrowserRouter>
 
-
-      </div>
     );
   }
 }
 
-export default mainPage;
+export default App;
