+++ frontend/src/components/ApprovalsTable.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import './tablestyles/globalTableStyles.css';
-import './tablestyles/approvalsTable.css';
-import Navbar from '../Navbar'
-import FiscalYear from '../FiscalYear';
+import '../styles/globaltablestyles.css';
+import '../styles/approvalstable.css';
+import Navbar from './Navbar'
+import FiscalYear from './FiscalYear';
 
 class ApprovalsTable extends Component {
   render() {
