+++ frontend/src/components/LeadsTable.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import './tablestyles/globalTableStyles.css';
-import './tablestyles/leads.css';
-import Navbar from '../Navbar'
-import FiscalYear from '../FiscalYear';
+import '../styles/globaltablestyles.css';
+import '../styles/leads.css';
+import Navbar from './Navbar'
+import FiscalYear from './FiscalYear';
 
 class LeadsTable extends Component {
     render() {
