+++ frontend/src/styles/globaltablestyles.css
