+++ frontend/src/styles/leaderboard.css
