+++ frontend/src/styles/leads.css
